requirements.txt
streamlit
ultralytics
Pillow
streamlit-webrtc
opencv-python-headless